+++ test/all.sh
@@ -13,5 +13,6 @@ $(dirname $0)/list.sh
 $(dirname $0)/show.sh
 $(dirname $0)/transform.sh
 $(dirname $0)/update.sh
+$(dirname $0)/generate.sh
 
 echo -e '\e[32mall tests passed!\e[0m'